move util files to util folder
