+++ src/gqlRules/utils/index.js
@@ -1,2 +1,5 @@
 export * from "./autoExport";
 export * from "./isGqlFile";
+export * from "./isGQLOperation";
+export * from "./operationName";
+export * from "./sanitizeGqlOperationText";
