+++ src/gqlRules/utils/isGQLOperation/index.js
