+++ src/gqlRules/utils/isGQLOperation/isQuery.js
