+++ src/gqlRules/utils/isGQLOperation/isQuery.spec.js
