+++ src/gqlRules/utils/sanitizeGqlOperationText/index.js
