+++ src/gqlRules/utils/sanitizeGqlOperationText/sanitizeGqlOperationText.js
