+++ src/gqlRules/utils/sanitizeGqlOperationText/sanitizeGqlOperationText.spec.js
